Organize header files
