+++ tooling/pipeline/includes/data_generation/prompt_formatting/prompt_formatter.h
@@ -1,4 +1,5 @@
-#pragma once
+#ifndef BIERGARTEN_PIPELINE_INCLUDES_DATA_GENERATION_PROMPT_FORMATTING_PROMPT_FORMATTER_H_
+#define BIERGARTEN_PIPELINE_INCLUDES_DATA_GENERATION_PROMPT_FORMATTING_PROMPT_FORMATTER_H_
 
 #include <string>
 #include <string_view>
@@ -15,3 +16,5 @@ class IPromptFormatter {
   [[nodiscard]] virtual std::string Format(
       std::string_view system_prompt, std::string_view user_prompt) const = 0;
 };
+
+#endif  // BIERGARTEN_PIPELINE_INCLUDES_DATA_GENERATION_PROMPT_FORMATTING_PROMPT_FORMATTER_H_
