+++ tooling/pipeline/includes/services/database/sqlite_export_service_helpers.h
@@ -1,0 +1,10 @@
+#ifndef BIERGARTEN_PIPELINE_INCLUDES_SERVICES_DATABASE_SQLITE_EXPORT_SERVICE_HELPERS_H_
+#define BIERGARTEN_PIPELINE_INCLUDES_SERVICES_DATABASE_SQLITE_EXPORT_SERVICE_HELPERS_H_
+
+/* Umbrella header for backward compatibility. */
+
+#include "sqlite_connection_helpers.h"
+#include "sqlite_handle_types.h"
+#include "sqlite_statement_helpers.h"
+
+#endif  // BIERGARTEN_PIPELINE_INCLUDES_SERVICES_DATABASE_SQLITE_EXPORT_SERVICE_HELPERS_H_
