+++ tooling/pipeline/includes/services/datetime/timer.h
@@ -1,5 +1,5 @@
-#ifndef BIERGARTEN_PIPELINE_INCLUDES_SERVICES_TIMER_H_
-#define BIERGARTEN_PIPELINE_INCLUDES_SERVICES_TIMER_H_
+#ifndef BIERGARTEN_PIPELINE_INCLUDES_SERVICES_DATETIME_TIMER_H_
+#define BIERGARTEN_PIPELINE_INCLUDES_SERVICES_DATETIME_TIMER_H_
 
 #include <chrono>
 
@@ -32,4 +32,4 @@ class Timer {
   }
 };
 
-#endif  // BIERGARTEN_PIPELINE_INCLUDES_SERVICES_TIMER_H_
+#endif  // BIERGARTEN_PIPELINE_INCLUDES_SERVICES_DATETIME_TIMER_H_
