--- tooling/pipeline/includes/services/sqlite_export_service_helpers.h
@@ -1,10 +1,0 @@
-#ifndef BIERGARTEN_PIPELINE_INCLUDES_SERVICES_SQLITE_EXPORT_SERVICE_HELPERS_H_
-#define BIERGARTEN_PIPELINE_INCLUDES_SERVICES_SQLITE_EXPORT_SERVICE_HELPERS_H_
-
-/* Umbrella header for backward compatibility. */
-
-#include "services/sqlite_connection_helpers.h"
-#include "services/sqlite_handle_types.h"
-#include "services/sqlite_statement_helpers.h"
-
-#endif  // BIERGARTEN_PIPELINE_INCLUDES_SERVICES_SQLITE_EXPORT_SERVICE_HELPERS_H_
