+++ tooling/pipeline/src/main.cc
@@ -21,13 +21,12 @@
 #include "data_generation/prompt_formatting/gemma4_jinja_prompt_formatter.h"
 #include "data_model/application_options.h"
 #include "llama_backend_state.h"
-#include "services/enrichment_service.h"
-#include "services/export_service.h"
-#include "services/prompt_directory.h"
-#include "services/sqlite_export_service.h"
-#include "services/timer.h"
-#include "services/wikipedia_service.h"
-#include "web_client/curl_web_client.h"
+#include "services/enrichment/enrichment_service.h"
+#include "services/database/export_service.h"
+#include "services/prompting/prompt_directory.h"
+#include "services/database/sqlite_export_service.h"
+#include "services/datetime/timer.h"
+#include "services/enrichment/wikipedia_service.h"
 #include "web_client/http_web_client.h"
 
 namespace di = boost::di;
