+++ tooling/pipeline/src/services/sqlite/finalize.cc
@@ -5,8 +5,8 @@
 
 #include <stdexcept>
 
-#include "services/sqlite_export_service.h"
-#include "services/sqlite_export_service_helpers.h"
+#include "services/database/sqlite_export_service.h"
+#include "services/database/sqlite_export_service_helpers.h"
 
 void SqliteExportService::Finalize() {
   if (db_handle_ == nullptr) {
