+++ tooling/pipeline/src/services/sqlite/initialize.cc
@@ -8,8 +8,8 @@
 #include <stdexcept>
 #include <string>
 
-#include "services/sqlite_export_service.h"
-#include "services/sqlite_export_service_helpers.h"
+#include "services/database/sqlite_export_service.h"
+#include "services/database/sqlite_export_service_helpers.h"
 
 std::filesystem::path SqliteExportService::BuildDatabasePath() const {
   std::filesystem::path base_filename("biergarten_seed_" + run_timestamp_utc_ +
